split jenkins-group-edit component (UI) (use tab)
