+++ angular-admin-ui/src/app/components/jenkins-group/edit/jenkins-group-edit.component.ts
@@ -5,8 +5,6 @@ import { Location }                 from '@angular/common';
 
 import { JenkinsGroup } from '../../../models/jenkins-group'
 import { JenkinsGroupService } from '../../../services/jenkins-group.service'
-import { JenkinsJob } from "../../../models/jenkins-job";
-import { JenkinsJobService } from "../../../services/jenkins-job.service";
 
 @Component({
   selector: 'app-jenkins-group-edit',
@@ -14,27 +12,18 @@ import { JenkinsJobService } from "../../../services/jenkins-job.service";
 })
 export class JenkinsGroupEditComponent implements OnInit {
   jenkinsGroup: JenkinsGroup;
-  jenkinsJobList: JenkinsJob[];
-  selectJenkinsJob: JenkinsJob;
-  refreshGrpahTrigger: number = 0;
+  test: number;
 
   constructor(
     private jenkinsGroupService:JenkinsGroupService,
-    private jenkinsJobServices: JenkinsJobService,
     private route: ActivatedRoute,
     private location: Location
-  ) {
-    this.selectJenkinsJob = new JenkinsJob();
-  }
+  ) { }
 
   ngOnInit() {
     this.route.paramMap
       .switchMap((params: ParamMap) => this.jenkinsGroupService.getJenkinsGroup(+params.get('id')))
       .subscribe(group => this.jenkinsGroup = group);
-
-    this.route.paramMap
-      .switchMap((params: ParamMap) => this.jenkinsJobServices.getJenkinsJobs(+params.get('id')))
-      .subscribe(jobs => this.jenkinsJobList = jobs);
   }
 
   save(): void {
@@ -42,49 +31,7 @@ export class JenkinsGroupEditComponent implements OnInit {
       .then(group => this.jenkinsGroup = group)
   }
 
-  editJob(job: JenkinsJob): void {
-    this.selectJenkinsJob = job;
-  }
-
-  saveJob(): void {
-    if (!this.selectJenkinsJob.id) {
-      this.selectJenkinsJob.jenkins_group_id = this.jenkinsGroup.id;
-      this.jenkinsJobServices.createJenkinsJob(this.selectJenkinsJob)
-        .then(job => {
-          this.jenkinsJobList.push(job);
-          this.selectJenkinsJob = new JenkinsJob();
-          this.refreshGraph();
-        })
-    } else {
-      this.jenkinsJobServices.updateJenkinsJob(this.selectJenkinsJob)
-        .then(job => {
-          this.selectJenkinsJob = job;
-          this.selectJenkinsJob = new JenkinsJob();
-          this.refreshGraph();
-        })
-    }
-  }
-
-  deleteJob(job: JenkinsJob): void {
-    if(confirm("Are you sure to delete job "+job.name)) {
-      this.jenkinsJobServices.deleteJenkinsJob(job)
-      .then(() => {
-        this.jenkinsJobList = this.jenkinsJobList.filter(j => j !== job);
-        this.refreshGraph();
-      });
-    }
-  }
-
-  updateWebHooks(): void {
-    this.jenkinsGroupService.updateJenkinsGroupWebhooks(this.jenkinsGroup.id)
-      .then(() => {})
-  }
-
   goBack(): void {
     this.location.back();
-  }
-
-  private refreshGraph() {
-    this.refreshGrpahTrigger += 1;
   }
 }
