+++ angular-admin-ui/src/app/components/jenkins-jobs/graph-widget/jenkins-jobs-graph-widget.components.css
