+++ angular-admin-ui/src/app/components/jenkins-jobs/jenkins-jobs.component.html
@@ -1,0 +1,104 @@
+<div class="col-xs-12 col-lg-12">
+  <section class="box">
+    <header>
+      <h2 class="title pull-left">jobs in group</h2>
+    </header>
+    <div class="content-body">
+      <ul class="nav nav-tabs">
+        <li [class.active]="tabSelected == 'jobs'"> <a (click)="selectTab('jobs')"  href="javascript:void(0)"> Jobs </a></li>
+        <li [class.active]="tabSelected == 'webhooks'"> <a (click)="selectTab('webhooks')"  href="javascript:void(0)"> Webhooks </a> </li>
+      </ul>
+      <div class="tab-content">
+        <!-- webhooks tab -->
+        <div class="tab-item top-buffer" *ngIf="tabSelected == 'webhooks'">
+          <a (click)="updateWebHooks()" class="btn btn-primary" href="javascript:void(0)"> Update WebHooks </a>
+          <table class="table table-hover top-buffer">
+            <thead>
+              <th>name</th>
+              <th>webhook url</th>
+              <th>actions</th>
+            </thead>
+            <tbody>
+              <tr *ngFor="let job of jenkinsJobList">
+                <td>{{job.name}}</td>
+                <td>http://host:port/gitlab/group/{{jenkinsGroup.name}}/job/{{ job.name }}</td>
+                <td>
+                  <a (click)="updateWebHooks()"  href="javascript:void(0)"> Update </a>
+                  <a (click)="updateWebHooks()"  href="javascript:void(0)"> Delete </a>
+                </td>
+              </tr>
+            </tbody>
+          </table>
+        </div>
+        <!-- jobs tab -->
+        <div class="tab-item top-buffer" *ngIf="tabSelected == 'jobs'">
+          <form class="form-inline" *ngIf="selectJenkinsJob" (ngSubmit)="f.form.valid && saveJob()" #f="ngForm">
+            <div class="form-group" [ngClass]="{ 'has-error': f.submitted && !name.valid }">
+              <input [(ngModel)]="selectJenkinsJob.name" [ngModelOptions]="{standalone: true}" #name="ngModel" class="form-control" placeholder="name" required>
+            </div>
+            <div class="form-group" [ngClass]="{ 'has-error': f.submitted && !gitLabProjectId.valid }">
+              <input [(ngModel)]="selectJenkinsJob.gitlab_project_id" [ngModelOptions]="{standalone: true}" #gitLabProjectId="ngModel" class="form-control" placeholder="gitlab_project_id" required>
+            </div>
+            <div class="form-group">
+              <select  [(ngModel)]="selectJenkinsJob.jenkins_job_perent_id" [ngModelOptions]="{standalone: true}" class="form-control" title="">
+                <option [ngValue]="null"></option>
+                <option *ngFor="let job of jenkinsJobList" [ngValue]="job.id">{{job.name}}</option>
+              </select>
+            </div>
+            <button class="btn btn-default">
+              <span class="glyphicon glyphicon-floppy-disk"></span>
+            </button>
+          </form>
+          <hr>
+          <table class="table table-hover">
+            <thead>
+              <tr>
+                <th>#</th>
+                <th>name</th>
+                <th>gitlab_project_id</th>
+                <th>jenkins_job_perent_id</th>
+                <th>actions</th>
+              </tr>
+            </thead>
+            <tbody>
+            <tr *ngFor="let job of jenkinsJobList">
+              <td>
+                <a (click)="editJob(job)"  href="javascript:void(0)">
+                  #{{job.id}}
+                </a>
+              </td>
+              <td>{{job.name}}</td>
+              <td>{{job.gitlab_project_id}}</td>
+              <td>{{job.jenkins_job_perent_id}}</td>
+              <td>
+                <a (click)="editJob(job)"  href="javascript:void(0)"> Edit </a>
+                <a (click)="deleteJob(job)" href="javascript:void(0)" > Delete </a>
+              </td>
+            </tr>
+            </tbody>
+          </table>
+        </div>
+      </div>
+    </div>
+  </section>
+</div>
+<div class="col-xs-12 col-lg-12">
+  <section *ngIf="jenkinsJobList" class="box">
+    <header>
+      <h2 class="title pull-left">
+        jobs graph
+      </h2>
+    </header>
+    <div class="content-body">
+      <div class="row">
+        <div class="col-xs-12 col-lg-12">
+          <p class="text-muted">Use the scroll to zoom</p>
+        </div>
+        <div class="col-lg-12 col-xs-12">
+          <app-jenkins-jobs-graph-widget [jenkinsJobList]="jenkinsJobList" [refreshGrpahTrigger]="refreshGrpahTrigger">
+          </app-jenkins-jobs-graph-widget>
+        </div>
+      </div>
+    </div>
+  </section>
+</div>
