+++ angular-admin-ui/src/index.html
@@ -6,7 +6,7 @@
   <base href="/ui/">
 
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="icon" type="image/x-icon" href="static/favicon.ico">
+  <link rel="icon" type="image/x-icon" href="/static/favicon.ico">
 </head>
 <body>
   <app-root></app-root>
