+++ angular-admin-ui/src/styles.css
@@ -18,6 +18,10 @@ body {
   border-bottom: 1px solid #eee;
 }
 
+.top-buffer {
+  margin-top:20px;
+}
+
 /*
  * Top navigation
  * Hide default border to remove 1px line.
